requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
iqoptionapi>=0.1.6
numpy>=1.24
pandas>=2.0
pandas-ta>=0.3.14b0
scikit-learn>=1.3
joblib>=1.3
python-machineid>=0.9.6
PyNaCl>=1.5
websocket-client>=1.6
pydantic>=2.0
